2nd Place/pyproject.toml
[project]
name = "snobert"
version = "0.1.0"
description = "Default template for PDM package"
dependencies = [
    "pandas>=2.2.1",
    "loguru>=0.7.2",
    "transformers>=4.38.2",
    "torch>=2.2.1",
    "scikit-learn>=1.4.1.post1",
    "timm>=0.9.16",
    "hydra-core>=1.3.2",
    "tensorboard>=2.16.2",
    "more-itertools>=10.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "jupyterlab>=4.1.4",
    "ipywidgets>=8.1.2",
]
